feat: Move ExtelResult from &dyn to Result<T, E>
ExtelResult originally was a dynamic trait object that restricted the
ability to use error propagation in tests. To fix this, ExtelResult now
is a type that represents a Result<(), Error>, allowing error
propagation as long as the Extel Error type supports it. Various doc
comments were updated to reflect this change.

Addresses: #7